+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [these versioning and changelog guidelines][changelog-gui
 
 ## Unreleased
 * Changed Action to Disableable Action in Card ([#838]https://github.com/Shopify/polaris-react/pull/838)
+### Enhancements
+* Added `renderChildren` prop to choice list component ([#993](https://github.com/Shopify/polaris-react/pull/993))
 
 ### Bug fixes
 * Fixed an issue with footer help links not expanding to full-width on mobile devices ([#759](https://github.com/Shopify/polaris-react/issues/759))